[feat] add ranges
This commit fully (well, not quite) ranges to the language. Ranges
are of two types, inclusive and exclusive. They can be created by
applying a range operator on two integers (.. and ..=). Applying
the range operator on anything else causes a panic. Under the hood,
ranges are represented by a struct that has the start, end and type
of the range.