[feat] add continue statement, remove ranges
This commit adds `continue` statements that can be used with loops.
A `continue` statement will move to the next iteration of the loop.
In case of a for loop, it will run the increment first, if it exists.
Using a `continue` outside a loop is a compile-time error.

Ranges have been removed, mainly because they weren't usable at all,
and seemed to be causing weird issues around string equality that I
couldn't figure out how to fix.