[feat] add break statements
This commit adds the commonly used break statements to zen. When the VM
encounters a `break`, it will immediately jump out of the nearest
enclosing loop. Trying to break outside a loop is a compile-time error.

All `Loop` structs now have a dynamic array of `break` offsets, which
consist of... well, offsets to break statements in the bytecode. This is
done since the jumps emitted by the break statement can't be patched
directly in the loop parsing functions using patch_jump(). Instead, a seperate
function handles patching these offsets one by one whenever a loop ends.