[feat] fix issue in switch statements with no else
This commit fixes the issue with switch statements mentioned in the
previous commit message. When a global variable was declared and then a
switch statement afterward had a block with a local variable declared in
it, that local variable would become equal to the global one. It was due
to weird stack pushes and pops, not quite sure how it happened myself.
Regardless, two changes to the switch statement managed to fix this
issue:

One, switch statements now must always have an `else` clause. With no
presence of strong typing in zen as of yet, this essentially makes them
exhaustive.

Two, the boolean comparison value left on the stack after cases being
compared with the switch value, and the switch value itself are popped
off the stack (just after the JUMP_IF_FALSE so as to not cause any
weirdness).

Doing this allows much more easier management of stack values in the
switch statement.