[feat] add functions!
Probably the biggest feature so far in zen! Functions are defined
with the `func` keyword, and act like any function you'd see in a
dynamically typed language. There is one weird limitation with them
though, in that a function cannot be recursive if it is not global.
At this point in Crafting Interpreters, Lox doesn't seem to support
those either, so I assume that comes later.

The availability of functions allows us to do a simple fibonacci
benchmark, using the also-added native function "clock" that returns
the Unix time in seconds. On benchmark, zen takes around 3.5 seconds to
generate the 35th fibonacci number. This is miles faster than Kaze,
which failed to generate the number in the first place. However, it is
still around 3.3 times slower than Python, which is what I aim to beat
for now.