[feat] add closures
This one took quite a bit of work, but closures are now present in
zen! They allow you to declare nested functions that can capture
variables of functions outside them, essentially carrying the
scope of their enclosing function even if that enclosing function
has long gone out of scope.

Closures in zen are implemented using upvalues. An upvalue simply is
a variable in an enclosing function. Whenever a function captures a
variable from an enclosing function, it creates an upvalue that
points to the location of the captured variable. If the upvalue is
still on the stack (an open upvalue) that location is simply the
index to it on the stack. If it has already gone off the stack, then
the location is a pointer to a heap allocation of the captured
variable.

Currently, closures seem to cause a lot of memory leaks, probably
since the GC hasn't been implemented yet. Still, I expected the
current object-freeing mechanism to take care of that. I'll have to
look into it, might be a mistake I made.