[feat] add anonymous and arrow functions, fix return bug
This commit adds two very useful features that make working with
programming functionally very easy: anonymous functions and arrow
notation for declaring functions.

Anonymous functions are simply functions without a name. Their syntax
is the exact same as that of a regular function, except the name is
removed. Whenever the parser sees the `func` keyword in an expression,
it parses the rest as a function. Anonymous functions are incredibly
useful since they make working with higher order functions very simple.

Another useful feature is arrow notation for function declaration.
These work the exact same way arrow functions do in JavaScript. After
the function's bracketed parameter list, one can choose to add a
`=>` instead of a brace. This causes the parser to parse whatever is
to the right of the arrow as an expression, and add a return instruction
just after that. This means that arrow functions return whatever the
expression after the `=>` is.

There has also been a bug fix regarding return statements. Previously,
if a function had the return statement in a block in the function, and
nowehere else, the implicit return that should have been added at the
end of the function would not be added, leading to segfaults when a
function like: `func f() { if false { return; } }` was called. This
issue has been fixed, and that code example now works fine.