[fix] remove memory leak caused by upvalue mismanagement
This commit fixes a memory leak caused by a rather dumb issue if I
say so myself.

So far, zen doesn't have a GC, so to free objects it relies on a
linked list of all the objects allocated, stored in the VM struct.
When the VM is freed, that list is traversed and every object is
freed one by one.

To allow for that, every Obj has a `next` field. This means that all
the "subtypes" of Obj like ObjString and ObjFunction, which "inherit"
Obj by using the `using` statement have a `next` field. The problem
is that upvalues also need a `next` field, so that upvalues that
link to other upvalues are possible. I named that field `next_upvalue`,
but I still managed to accidentally use `next` where I should have
used `next_upvalue`; this resulted in disastrous consequences where
the link between upvalues and the rest of the Objs in the linked list
of Objs was broken, and caused a lot of objects to leak in memory.

Thankfully due to Odin's tracking allocator, I was able to find out
that memory was leaking immediately. It took some time before I
realized that it was due to my typos, though.