[refactor] replace just with python
This commit replaces the current justfile-based build system with a Python script.
Just was far less verbose and very easy to work with, but Python has much better cross
platform compatibility and it just works anywhere; additionally it has a powerful
argument parser that helps a lot. I might change this decision later on, but for now,
Python script it is.