[feat] add pipelines
Big feature! This feature allows convenient and very powerful piping of expressions.
The pipe operator "|>" is a binary operator that has two expressions as its operands.
When a pipe operator is encountered after an expression, it means "keep the result of
the expression just compiled on the stack, to be used by the next expression". The
piped value can be used in two ways.

The first method is function calls. When an expression is piped into a function call,
that expression is inserted as the first argument of that call. So, a |> f(b) is
equivalent to f(a, b). Additionally, a |> f(b) |> g(c) is also equivalent to
g(f(a), c). This mechanism allows for very elegant-looking code with a functional
flavor to it.

The other way to use the piped value is via the newly added "it" variable. "it" is a
variable whose value is equal to the previously piped value. You can use "it" in cases
where function calls won't do what you want. For example,
i |> it + 2 |> it * 3 |> it / 9 |> it - 2; is equivalent to:
(((i + 2) * 3) / 9) - 2;

It looks more verbose, but splitting it in seperate lines is wonderful for readability.
Using "it" outside of a pipeline results in a compilation error.

Other than the pipes, a new native function "reverse()" that reverses a string passed
to it has been added.