[feat] add complex numbers and typeof
This commit adds the 'complex' datatype that represents any
complex number. Complex numbers can be created using the native
function complex(r, i). They support all four basic arithmetic
operations, and there's also a couple of functions to help out with
them, like conjg(z) to get the complex conjugate, and abs(z) to get
the magnitude (or the absolute value if the argument is real).

Additionally, this commit also adds the typeof(x) native function
which returns the type of any expression, as a string.