[feat] add lists
Another big feature lands! This time, we have lists. Finally, a
data structure! This is massive because the list makes it possible
to do so many things that would have been extremely impractical
and difficult without it.

Lists are very basic here; you create one with a list literal
using square brackets, you can subscript them with the square
bracket syntax too, and the push() and pop() functions are
available to add and remove items from a list. Quite simple.