[feat] add methods, initializers and `this`
This commit makes classes a whole lot useful, adding in methods
that you can define in the class body, as well as the initializer
which is a special method that runs when an instance is created,
along with the `this` keyword used to access the instance when in
a method.

Syntax is described in the documentation file.