[fix] add missing line of code in values_equal()
In the earlier commit I said over 50 tests were failing. Turns out
that was because I missed a return statement after an if statement,
and if that statement evaluated to false the program would reach
a call to `unreachable()` which was causing SIGILLs making the
tests fail. I kinda wish Odin would say something like "You told me
this code was unreachable you idiot" instead of an ambigious error,
but then again it's mainly my stupidity anyways.

Problem fixed now though, and all the tests are passing!