feat: add basic framework for user-defined imports
This basically works by running the imported program when the importer's
VM is running by way of a new VM. I've also added protection against
cyclic imports using a basic import stack managed by the GC.

Also, the reason why the imported programs require information of the
importer is due to the fact that the importer's VM's stack stores the
module that will keep all of the global functions in the imported file.
That functionality is yet to be added.