fix: failing superclass invocation test
Involved changing `vm.stack` to be a constant array rather than a
dynamic array, as whenever the dynamic array was updated in size, the
`frame.slots` pointer in the current callframe had would get corrupted.
Using a constant array eliminates that issue entirely. The one problem
that does arise is much higher memory usage, but that is no more than a
few hundred kilobytes.